use require instead of namespacing
