+++ README.md
@@ -142,3 +142,8 @@ $result = batch()->update($userInstance, $value, $index);
 
 $result = batch()->insert($userInstance, $columns, $values, $batchSize);
 ```
+
+# Tests
+If you don't have phpunit installed on your project, first run `composser require phpunit/phpunit`
+
+In the root of your laravel app, run `./vendor/bin/phpunit ./vendor/mavinoo/laravel-batch/tests`
