+++ tests/TestModel.php
@@ -1,0 +1,11 @@
+<?php
+
+class TestModel extends \Illuminate\Database\Eloquent\Model
+{
+    protected $table = 'HmKTjCJgLN9bBq7KXzI3';
+
+    public function tableName()
+    {
+        return $this->table;
+    }
+}
